// src/thirdparty/Genshin-Impact-Miliastra-Wonderland-Code-Node-Editor-Pack/protobuf/gia.proto.ts
export const VarType = {
  UnknownVar: 0,
  Entity: 1,
  GUID: 2,
  Integer: 3,
  Boolean: 4,
  Float: 5,
  String: 6,
  GUIDList: 7,
  IntegerList: 8,
  BooleanList: 9,
  FloatList: 10,
  StringList: 11,
  Vector: 12,
  EntityList: 13,
  EnumItem: 14,
  VectorList: 15,
  LocalVariable: 16,
  Faction: 17,
  Configuration: 20,
  Prefab: 21,
  ConfigurationList: 22,
  PrefabList: 23,
  FactionList: 24,
  Struct: 25,
  StructList: 26,
  Dictionary: 27,
  VariableSnapshot: 28
} as const
export type VarType = (typeof VarType)[keyof typeof VarType]
export const ClientVarType = {
  UnknownVar_: 0,
  Entity_: 1,
  EntityList_: 2,
  Integer_: 3,
  IntegerList_: 4,
  Boolean_: 5,
  BooleanList_: 6,
  Float_: 7,
  FloatList_: 8,
  String_: 9,
  StringList_: 10,
  Vector_: 11,
  VectorList_: 12,
  EnumItem_: 13,
  GUID_: 14,
  GUIDList_: 15,
  Faction_: 16,
  EnumList_: 17,
  Configuration_: 18,
  Prefab_: 19,
  ConfigurationList_: 20,
  PrefabList_: 21,
  LocalVariable_: 22
} as const
export type ClientVarType = (typeof ClientVarType)[keyof typeof ClientVarType]